Cleanup and refactor
