--- scripts/app/configs/paths.ts
@@ -1,17 +1,0 @@
-export interface Paths {
-    build: string;
-    config: string;
-    modules: string;
-    public: string;
-    src: string;
-}
-
-const paths: Paths = {
-    build: "build",
-    config: "config",
-    modules: "node_modules",
-    public: "public",
-    src: "src",
-};
-
-export default paths;
